Updated README & replaced main.py with example.py
