+++ README.md
@@ -1,8 +1,15 @@
 # revoltapilibrarything
+
 Python wrapper for the revolt api ~~like we didnt already have enough of them~~
+
+
 ## Why make another wrapper?
+
 Because I was bored, and I needed something to turn in as a final project for my programming class
+
+
 ## Features
+
 - Websocket events
 - More manual than others, no command processor
 - Server fetching
@@ -10,10 +17,21 @@ Because I was bored, and I needed something to turn in as a final project for my
 - Message fetching(not by id)
 - Channel fetching
 - Member fetching
-- Client class wrapping everything else, meaning 1 import
+- Client class wrapping everything else, meaning 1 import only (`import client`)
 - Message sending
 - Support for both user and bot tokens
-### Documentation
+
+
+## Getting Started
+
+To get started using the API, you have to use `git clone https://github.com/error-404-null-not-found/revoltapilibrarything` and edit the example.py file which shows some of the API's basic functionality.
+
+
+## Documentation
+
 Documentation can be found [here](https://error-404-null-not-found.github.io/revoltapilibrarything/Documentation/index.html)
-### Support
+
+
+## Support
+
 Join the support server on [Revolt](https://revolt.chat): https://app.revolt.chat/invite/vkzrnSgt
